Beta

+++ sounds.ino
@@ -1,4 +1,4 @@
-const int soundfx[6][8] = {
+const byte soundfx[6][8] = {
   {1,17,53,0,7,0,2,3}, // 0 = thrust (channel 0)          fire
   {1,17,53,0,7,0,10,3}, // 0 = thrust (channel 0)          fire
   {1,26,41,1,1,3,7,20}, // 1 = crash (channel 1)          boom
